langgraph/requirements.txt
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-anthropic>=0.3.0
snowflake-connector-python>=3.0.0
